Added save gui
There is now a gui that has the functionality to select a file to save to.
There is also a new Flashcard class that holds the term and definition of a flashcard.
There are also some new variables including a FIle that holds the selected save file, two arraylists that hold flashcards for two different guis, and a JFIleChooser for the save gui.