Added various features
The two guis will display the currently selected files. The save gui will display cards that are added to the list. The progam will add cards saved to the file  to the displayCards list only if the files selected on both guis have the same path to keep the list of cards updated.  Added a check to make sure that a file is selected before saving to anything.